# API.yml
openapi: 3.0.0
info:
  title: Mobile Number API
  description: ""
  version: 1.0.0
servers:
  - url: http://localhost:8080/api/v1
    description: Temporary host at version 1

paths:
  /mobile-numbers:
    get:
      summary: Returns a list of subscribed mobile numbers.
      description: Returns a list of mobile number subscriptions contained in the database.
      responses:
        200:
          description: A JSON array of mobile number subscriptions
          content:
            application/json:
              schema: 
                type: array
                items:
                  $ref: '#/components/schemas/MobileNumber'
        default:
          description: Common Unexpected Errors
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
    post:
      summary: Add a new mobile number subscription
      description: "Add a new Subscriber with unique number mobile"
      requestBody:
        content:
          application/json:
            schema:
                $ref: '#/components/schemas/MobileNumber'
              
      responses:
        201:
          description: Mobile number subscription
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MobileNumber'
        
        400:
          $ref: '#/components/responses/BadRequest'

        401:
          $ref: '#/components/responses/Unauthorized'
        
        default:
          description: Common Unexpected Errors
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'

            
  /mobile-numbers/{id}:
    parameters:
    - name: id
      in: path
      description: "ID of mobile number subscription to return"
      required: true
      schema:
        type: integer
        format: int64
    get:
      summary: Returns mobile number subscription
      description: "Subscriber by ID"
      responses:
        200:
          description: Subscriber Details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MobileNumber'
        
        400:
          $ref: '#/components/responses/BadRequest'

        401:
          $ref: '#/components/responses/Unauthorized'
          
        404:
          $ref: '#/components/responses/NotFound'
        
        default:
          description: Common Unexpected Errors
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
                
    delete:
      summary: Delete a mobile number subscription
      description: "Delte mobile number subscription by the given ID"
      responses:
        204:
          description: Subscriber Deleted Correctly

        401:
          $ref: '#/components/responses/Unauthorized'
        
        default:
          description: Common Unexpected Errors
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
    patch:
      summary: Patch mobile number subscription details
      description: "Patch mobile number subscription details by its ID"
      responses:
        200:
          description: Subscriber Details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MobileNumber'
        
        400:
          $ref: '#/components/responses/BadRequest'

        401:
          $ref: '#/components/responses/Unauthorized'
          
        404:
          $ref: '#/components/responses/NotFound'
        
        default:
          description: Common Unexpected Errors
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'

components:
  schemas:
    MobileNumber:
      type: object
      properties:
        id:
          type: integer
          format: int64
        mobileNumber:
          type: string
          pattern: '^[1-9]\d{1,14}$'
          example: "35699123456"
        userId:
          type: integer
          format: int64
        ownerId:
          type: integer
          format: int64
        serviceType:
          type: string
          enum: ["Prepaid", "Postpaied"]
        subscribtionDate:
          type: string
          format: date-time
          
    ErrorMessage:
      type: object
      properties:
        timestamp: 
          type: string
        status:
          type: integer
        error: 
          type: string
        message: 
          type: string
        path: 
          type: string

  responses:
    Unauthorized:
      description: Unauthorised Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorMessage'
            
    NotFound:
      description: Resource Not Found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorMessage'
            
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorMessage'
            
    UnexpectedError:
      description: Unexpected Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorMessage'
        